requirements.txt
flask==1.1.2
flask-restx==0.2.0
flask-sqlalchemy==2.4.4
flask-jwt-extended==4.0.2
flask-migrate==2.7.0
python-dotenv==0.15.0
werkzeug==1.0.1
gunicorn==20.0.4